Style/organization cleanup
